удаление ненужного2
