--- sort_array/sort_array/Program.cs
@@ -1,51 +1,0 @@
-﻿
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace SortArray
-
-{
-    public class Program
-    {
-        /// <summary>
-        /// эта штука сортирует массив
-        ///
-        /// </summary>
-        /// 
-        public static void Main(string[] args)
-        {
-            Console.WriteLine("Enter amount");
-            int number = int.Parse(Console.ReadLine());
-            int[] arr = new int[number];
-            Console.WriteLine("Enter numbers");
-            string String = Console.ReadLine();
-            int count = 0;
-            foreach (int v in String.Split(' ').Select(v => Convert.ToInt32(v)))
-            {
-                arr[count] = v;
-                count++;
-            }
-            for (count = 0; count < number; count++)
-            {
-                for (int count2 = 0; count2 < number - 1; count2++)
-                {
-                    if (arr[count2] > arr[count2 + 1])
-                    {
-                        int t = arr[count2];
-                        arr[count2] = arr[count2 + 1];
-                        arr[count2 + 1] = t;
-                    }
-                }
-            }
-            for (count = 0; count < number; count++)
-            {
-                Console.Write($"{arr[count]} ");
-            }
-
-            Console.ReadKey();
-        }
-    }
-}
